Move view-related types and extensions to the UserInterface package
