+++ Sources/Player/DequeuePlayer.swift
@@ -8,8 +8,8 @@ import AVFoundation
 
 // TODO: Must later implement dequeue support for items to have playlist support in forward and backward directions.
 @MainActor
-final class DequeuePlayer: AVQueuePlayer {
-    override func seek(to time: CMTime, toleranceBefore: CMTime, toleranceAfter: CMTime, completionHandler: @escaping (Bool) -> Void) {
+public final class DequeuePlayer: AVQueuePlayer {
+    public override func seek(to time: CMTime, toleranceBefore: CMTime, toleranceAfter: CMTime, completionHandler: @escaping (Bool) -> Void) {
         NotificationCenter.default.post(name: .willSeek, object: self, userInfo: [
             SeekInfoKey.targetTime: time
         ])
