+++ Sources/Player/Extensions.swift
@@ -6,7 +6,6 @@
 
 import Combine
 import Foundation
-import SwiftUI
 
 extension Comparable {
     func clamped(to range: ClosedRange<Self>) -> Self {
@@ -65,28 +64,3 @@ public extension Publisher {
         weakCapture(other, at: \T.self)
     }
  }
-
-public extension Binding {
-    /// Create a binding to an object property.
-    /// - Parameters:
-    ///   - object: The object to bind to.
-    ///   - keyPath: The key path to bind to.
-    init<T>(_ object: T, at keyPath: ReferenceWritableKeyPath<T, Value>) {
-        self.init(
-            get: { object[keyPath: keyPath] },
-            set: { object[keyPath: keyPath] = $0 }
-        )
-    }
-
-    /// Create a binding to an object nullable property.
-    /// - Parameters:
-    ///   - object: The object to bind to.
-    ///   - keyPath: The key path to bind to.
-    ///   - defaultValue: The default value to use when the property is `nil`.
-    init<T>(_ object: T, at keyPath: ReferenceWritableKeyPath<T, Value?>, defaultValue: Value) {
-        self.init(
-            get: { object[keyPath: keyPath] ?? defaultValue },
-            set: { object[keyPath: keyPath] = $0 }
-        )
-    }
-}
