+++ Sources/UserInterface/Slider.swift
@@ -4,6 +4,7 @@
 //  License information is available from the LICENSE file.
 //
 
+import Player
 import SwiftUI
 
 @available(tvOS, unavailable)
