--- Sources/UserInterface/UserInterface.swift
@@ -1,9 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-/// User interface
-public struct UserInterface {
-}
